+++ src/options.h
@@ -341,7 +341,7 @@ struct options_main {
 	int max_run_time;
 
 /* Graceful exit after this many candidates tried. */
-	int64_t max_cands;
+	unsigned long long max_cands;
 
 /* Emit a status line every N seconds */
 	int status_interval;
